requirements.txt
djangorestframework==3.12.2
Django==3.1.2
psycopg2==2.8.6
django-heroku==0.3.1
gunicorn==20.0.4
django-cors-headers==3.2.1
python-decouple==3.3